:white_check_mark: Fix reproducibility tests for ChemTorch benchmark - Fix val loss parsing in extraction helpers - Update baselines and reference predictions. - Fix tests to run on CPU for consistency
NOTE: The original experiments were run on an A40 GPU but tests were executed on a CPU (e.g. laptop or CI/CD). This lead to test failures in the due to slight numerical inaccuracies in test predictions and the 3-epoch val loss. Consequently, reproducibility was verified manually for the latest ChemTorch version on an A40 GPU for the seed 0 runs (as in the original experiments). Subsequently, the test configurations were modified to run on CPU to ensure consistency across different environments, and the baselines and reference predictions were updated accordingly.